Add asha support with stub set configuration
This commit shows the direction the asha implementation can take if we
make it depend on the media implementation (and by extension, for now
the a2dp implementation) in profiles/audio.

Since currently the media structures (media_endpoint, media_transport
and so on) have a2dp fields, we will add similar asha fields as
necessary. A lot of code will have to deal with both branches.

A future refactor can address the coupling between the media interfaces
and the a2dp/asha implementions. Merging the asha changes can also be
contingent on this refactor being complete.